Create separate entity RejectionReason.
