+++ huntflow_api_client/models/response/applicant_on_vacancy.py
@@ -26,18 +26,6 @@ class VacancyStatusesResponse(BaseModel):
     items: List[VacancyStatus]
 
 
-class RejectionReason(BaseModel):
-    """The ID can be None because the backend adds the rejection reason "Other" where ID is None"""
-
-    id: Optional[int] = Field(None, description="Rejection reason ID")
-    name: str = Field(..., description="Rejection reason name")
-    order: int = Field(..., description="Order")
-
-
-class RejectionReasonsListResponse(BaseModel):
-    items: List[RejectionReason]
-
-
 class AddApplicantToVacancyResponse(BaseModel):
     id: int = Field(..., description="Binding ID", example=9)
     changed: datetime = Field(..., description="The date of recruitment stage change")
